android_aws_device_farm/requirements.txt
pytest==3.0.7
Appium-Python-Client==0.24
selenium==3.0.0
six==1.10.0
py==1.4.33